2930 - Sanitize sorting and comparisons
Change 1: added missing less-than method for sorting
The ScriptResult class in nmap/ndiff/ndiff.py is not specifying a __lt__ (less-than) method by which the precedence of its instances could be compared when sorting.
This can lead to an error on the sort method call in line 1420.
self.current_port.script_results.sort()
The error occurs when multiple script results are exported to xml and parsed by the underlying xml.sax library.

Change 2: added missing len function call when comparing list to int in conditional statement
Conditional statement in line 983 throws a TypeError when script results have differences for the same port.

Change 3: implemented host object sorting and comparative logic  in OO style
The Host class does now define the equals and less-than methods. This allows for the get_id() method of the Host class to be treated as private and comparisons to be made directly against between instances.
The `Scan` objects list of addresses is not being mutated in-place, but instead wrapped in a `sorted` function call and passed to `host_pairs()` in line 552.


Change 4: Removed OO boilerplate
Python3 classes inherit from the object base class implicitly, therefore `class A (object)` can be reduced to `class A`.
See accepted answer in https://stackoverflow.com/questions/4015417/why-do-python-classes-inherit-object and 

In Python3 the not-equals method implicitly delegates to the equals method. Therefore it is unnecessary to explicitly define `__ne__`  as the negation of the corresponding `__eq__` method. See https://docs.python.org/3/reference/datamodel.html#object.__eq__