Fix passing argument 4 of ReadFile from incompatible pointer type
ReadFile returns the number of read bytes in a DWORD, i.e. unsigned long,
not in an int. GCC 14 throws an incompatible pointer type due to this,
at least in Debian.

This patch corrects the type of variable bytes_read to DWORD and mangles the
returned value into an int when it is used with calling output(), it
cannot be greater than 1023 (= 0x3FF) anyway.