Flag to allow sqlping payloads in non service and version detection scans
We (Rapid7) would like to move to using upstream nmap. We require
payloads to be sent with Sqlpings in non service and version detection
scans. Currently we need to apply our own patch to nmap-service-probes
in order to acheieve this behaviour. This change adds this behaviour to
upstream nmap but can only be enabled using the `--sqlping-payload` flag